Add files via upload
We often include conditionals for when we need to account for different situations and outcomes. When it comes to coming up with a plan and executing it, there’s nothing more uncertain than a bank heist! (Watch any robbery themed movie if you need convincing).

In this project, we’re going to use conditionals to simulate this situation along with hi-jinks and hiccups that may pop up. Who knows? Maybe, just maybe, if all goes well, we’ll have a few more dollars after.